The Last Night Push
This was the stuff I had worked on for Goblin Boxer overnight. I had been up until like 4 in the morning working on this and had to stop myself from pulling an all nighter working on this. I will do better on this front in the case of managing my time better so that stuff like this doesn't happen in the future.